requirements.txt
pyrogram
pytesseract
Pillow
tgcrypto
python-dotenv
autocorrect
google-generativeai 